+++ src/main/java/org/folio/circulation/loanrules/ItemType.java
@@ -1,10 +1,21 @@
 package org.folio.circulation.loanrules;
 
+/**
+ * Store the UUID of the item's type (the material type).
+ */
 public class ItemType {
+  /** UUID of the item type. */
+  @SuppressWarnings("squid:ClassVariableVisibilityCheck")  // Drools directly uses public fields
   public String id;
+
+  /**
+   * Set item type.
+   * @param id  UUID of the item type
+   */
   public ItemType(String id) {
     this.id = id;
   }
+
   @Override
   public String toString() {
     return id;